remove various compiler warnings

+++ geda/graphman/configfile.c
@@ -127,7 +127,7 @@ int ConfigGetNext(char *szName, char *szValue)
 			continue;
 		if (szLine[0] == '[')
 			break;
-		if (isascii(szLine[0]))
+		if (isascii((int) szLine[0]))
 			iFound = 1;
 		break;
 	}
@@ -135,10 +135,10 @@ int ConfigGetNext(char *szName, char *szValue)
 	if (iFound == 0)
 		return FAILURE;
 
-	for (i = 0; i < strlen(szLine) && isalnum(szLine[i]); i ++)
+	for (i = 0; i < strlen(szLine) && isalnum((int) szLine[i]); i ++)
 		szName[i] = toupper(szLine[i]);
 	szName[i] = 0;
-	for (; i < strlen(szLine) && (szLine[i] == '=' || isspace(szLine[i])); i ++)
+	for (; i < strlen(szLine) && (szLine[i] == '=' || isspace((int) szLine[i])); i ++)
 		;
 	for (; i < strlen(szLine); i ++)
 		szValue[j ++] = szLine[i];
@@ -187,7 +187,7 @@ char *ConfigValue(const int iValue)
 			switch (iState)
 			{
 				case PARSE_SPACES:
-					if (!isspace(szLineBuf[i]))
+					if (!isspace((int) szLineBuf[i]))
 					{
 						if (szLineBuf[i] == '"')
 						{
@@ -217,7 +217,7 @@ char *ConfigValue(const int iValue)
 					break;
 
 				case PARSE_ALNUM:
-					if (isspace(szLineBuf[i]) || IsSeparator(szLineBuf[i]))
+					if (isspace((int) szLineBuf[i]) || IsSeparator(szLineBuf[i]))
 					{
 						iState = PARSE_STOP;
 					}
